fixed the look of welcome screen and resized the items proportional to screen
I fixed the design of the welcome screen according to the figma design.

since the hocam connect logo is overflowing in the figma design, i encountered a problem, i cropped the image but i eyeballed it.
wrote a screen config file that can be used a template to create every screen items proportional to the screen size. Fixed only the welcome screen according to it.
*Will fix the other screens later
*I'm not even sure if this is needed since i couldn't find anything abour it online.